requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
et-xmlfile==1.1.0
# gevent==21.8.0
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
kombu==5.1.0
lml==0.1.0
openpyxl==3.0.9
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
pyexcel-io==0.6.4
pyexcel-xlsx==0.6.0
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
sib-api-v3-sdk==7.2.0
six==1.16.0
sqlparse==0.4.1
ua-parser==0.10.0
urllib3==1.26.6
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
zope.event==4.5.0
zope.interface==5.4.0
django-tenants-celery-beat
django-celery-results>=2.2.0,<2.3.0
django-multiselectfield>=0.1.12,<0.2.0
pysimilar>=0.4,<1.0
textblob>=0.17.1,<0.18.0
pdfminer.six>=20211012
Pillow>=9.0.1,<9.2.0
black>=21.8b0,<21.9b0
Django>=3.2.5,<3.3.0
django-cors-headers>=3.7.0,<3.8.0
django-environ>=0.4.5,<0.5.0
django-filter>=2.4.0,<2.5.0
django-redis>=5.0.0,<5.1.0
django-tenants>=3.3.2,<3.4.0
django-user-agents>=0.4.0,<0.5.0
djangorestframework>=3.12.4,<3.13.0
djangorestframework-simplejwt>=4.7.2,<4.8.0
psycopg2-binary>=2.9.1,<2.10.0
django-silk>=4.2.0,<4.3.0
cloudinary>=1.29.0,<1.30.0
dj3-cloudinary-storage>=0.0.6,<0.1.0
django-cleanup>=6.0.0<=6.1.0
PyYAML>=6.0,<7.0
django-import-export>=2.8.0,<2.9.0
